Add XEP revision details.
For each implemented XEP, this:

* adds a column with the title of the XEP,
* the XEP abstract is injected in the "title" attribute;
* adds a column with the latest revision number,
* the date of the latest revision is injected in the "title"
  attribute;
* adds a "title" attribute to the implemented revision number
  with the publication date of that version; and
* styles the implemented revision column according to whether
  the project implements a current or outdated revision.

It does this by referencing the relevant XEP XMLs, not the
abridged list found at https://xmpp.org/extensions/xeplist.xml
as that one does not have full revision details (so we can extract
the date of the *implemented*, as opposed to latest, revision) and
it also makes matching marginally more awkward as the XEP numbers
are given with a varying number of digits. Lastly, that relies on
xeplist.xml being complete and up to date whereas taking the XEP
references entered by the user relies on the *.html and *.xml
versions being in the same path. One marginal advantage is that
this doesn't break if the author announces support for private
XMLs (which probably won't be in xmpp.org or in xeplist.xml).

Note 1: Generation with xsltproc will probably not work:
https://gitlab.gnome.org/GNOME/libxslt/-/issues/28 Use a better
tool instead, such as Saxon (http://saxon.sourceforge.net/).

Note 2: If the XEP XMLs were served with
`Access-Control-Allow-Origin: *`, the XML could be served directly
and the browser would apply the XSLT transformation itself.